refactor: split cli into multiple files
